support read schema
fix style:

fix style
