+++ core/src/main/java/org/apache/carbondata/core/util/CarbonUtil.java
@@ -86,8 +86,6 @@ import org.apache.carbondata.core.util.path.CarbonTablePath;
 import org.apache.carbondata.format.BlockletHeader;
 import org.apache.carbondata.format.DataChunk2;
 import org.apache.carbondata.format.DataChunk3;
-import org.apache.carbondata.format.FileHeader;
-
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
@@ -2336,19 +2334,12 @@ public final class CarbonUtil {
   /**
    * This method will read the schema file from a given path
    *
-   * @param schemaFilePath
-   * @return
+   * @return table info containing the schema
    */
   public static org.apache.carbondata.format.TableInfo inferSchema(
-      String carbonDataFilePath, AbsoluteTableIdentifier absoluteTableIdentifier,
+      String carbonDataFilePath, String tableName,
       boolean schemaExists) throws IOException {
-    TBaseCreator createTBase = new ThriftReader.TBaseCreator() {
-      public org.apache.thrift.TBase<org.apache.carbondata.format.TableInfo,
-          org.apache.carbondata.format.TableInfo._Fields> create() {
-        return new org.apache.carbondata.format.TableInfo();
-      }
-    };
-    if (schemaExists == false) {
+    if (!schemaExists) {
       List<String> filePaths =
           getFilePathExternalFilePath(carbonDataFilePath + "/Fact/Part0/Segment_null");
       String fistFilePath = null;
@@ -2358,16 +2349,9 @@ public final class CarbonUtil {
         LOGGER.error("CarbonData file is not present in the table location");
       }
       CarbonHeaderReader carbonHeaderReader = new CarbonHeaderReader(fistFilePath);
-      FileHeader fileHeader = carbonHeaderReader.readHeader();
-      List<ColumnSchema> columnSchemaList = new ArrayList<ColumnSchema>();
-      List<org.apache.carbondata.format.ColumnSchema> table_columns = fileHeader.getColumn_schema();
-      for (int i = 0; i < table_columns.size(); i++) {
-        ColumnSchema col = thriftColumnSchmeaToWrapperColumnSchema(table_columns.get(i));
-        col.setColumnReferenceId(col.getColumnUniqueId());
-        columnSchemaList.add(col);
-      }
+      List<ColumnSchema> columnSchemaList = carbonHeaderReader.readSchema();
       TableSchema tableSchema = new TableSchema();
-      tableSchema.setTableName(absoluteTableIdentifier.getTableName());
+      tableSchema.setTableName(tableName);
       tableSchema.setBucketingInfo(null);
       tableSchema.setSchemaEvalution(null);
       tableSchema.setTableId(UUID.randomUUID().toString());
@@ -2392,6 +2376,12 @@ public final class CarbonUtil {
       tableInfo.setDataMapSchemas(null);
       return tableInfo;
     } else {
+      TBaseCreator createTBase = new ThriftReader.TBaseCreator() {
+        public org.apache.thrift.TBase<org.apache.carbondata.format.TableInfo,
+            org.apache.carbondata.format.TableInfo._Fields> create() {
+          return new org.apache.carbondata.format.TableInfo();
+        }
+      };
       ThriftReader thriftReader = new ThriftReader(carbonDataFilePath, createTBase);
       thriftReader.open();
       org.apache.carbondata.format.TableInfo tableInfo =
